backend/requirements.txt
aiosmtplib==4.0.1
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.0.1
black==25.12.0
certifi==2025.8.3
cffi==1.17.1
cfgv==3.5.0
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.5
Deprecated==1.2.18
distlib==0.4.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.13
filelock==3.20.1
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.15
idna==3.10
iniconfig==2.1.0
isort==7.0.0
librt==0.7.5
limits==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
psycopg2==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.12.0
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-asyncio==1.1.0
python-dotenv==1.1.0
python-jose==3.5.0
pytokens==0.3.0
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
ruff==0.14.10
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
structlog==25.5.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
virtualenv==20.35.4
wrapt==1.17.2
